plug camera in map scene and adjust settings
