+++ Assets/Prefabs/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fa0d638f39a42184b9ea6080cd7a8058
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
