+++ Assets/Prefabs/Player/PlayerCameraController.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 0e60e097f7970aa4dbabf2fe9573b201
-DefaultImporter:
+guid: abc5367a666300b4cb949d54f804272e
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
