+++ Assets/Scenes/Map/MapScene.unity
@@ -135,10 +135,10 @@ GameObject:
   - component: {fileID: 151935713}
   - component: {fileID: 151935712}
   - component: {fileID: 151935711}
-  - component: {fileID: 151935710}
   - component: {fileID: 151935709}
+  - component: {fileID: 151935710}
   m_Layer: 8
-  m_Name: Main Camera
+  m_Name: Camera
   m_TagString: MainCamera
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -156,7 +156,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 3
+  m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 20, y: 0, z: 0}
 --- !u!114 &151935709
 MonoBehaviour:
@@ -192,7 +192,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _target: {fileID: 0}
   zoom: 0
-  distance: 60
+  distance: 30
   TargetRotation: 0
   Direction: 0
   Altitude: 35
@@ -364,7 +364,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 0
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1169955387
 GameObject:
@@ -395,7 +395,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1169955389
 MonoBehaviour:
@@ -411,6 +411,67 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   worldLight: {fileID: 2131324329}
   UIController: {fileID: 0}
+--- !u!1001 &1998137136
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_RootOrder
+      value: 5
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683012, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2714124424983683013, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: GameCamera
+      value: 
+      objectReference: {fileID: 151935713}
+    - target: {fileID: 2714124424983683014, guid: abc5367a666300b4cb949d54f804272e, type: 3}
+      propertyPath: m_Name
+      value: PlayerCameraController
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: abc5367a666300b4cb949d54f804272e, type: 3}
 --- !u!1 &2131324328
 GameObject:
   m_ObjectHideFlags: 0
@@ -501,5 +562,5 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 1
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
