+++ Assets/Scripts/GameCamera.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8854970581030f446ab11683b06cc6eb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
